backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0
structlog==24.1.0
aioboto3==12.3.0
pyarrow==15.0.0
pandas==2.2.0
chardet==5.2.0
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
slowapi==0.1.9
python-multipart==0.0.22

# Dev dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
respx==0.20.2
moto==5.0.0
ruff==0.1.9
mypy==1.8.0
